[NHDR-198] fix: 기존 파일 참고하여 수정
